refactor: make destructurization of path validation
